Automatically merge branch "new_reelase"
This branch is intended to be used for a pipeline that automatically
creates new Docker images when a new StackMaster version is released.
Part of the necessary steps will be that a commit with the version
update is pushed to a branch "new_release". After the correctness of the
new branch is checked we can automatically merge it to master.